Correções linters
