+++ components/ui/multi-select.tsx
@@ -2,9 +2,7 @@
 
 import * as React from 'react'
 
-// Simplified version without Popover - imports removed to avoid unused warnings
-// import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
-// import { MultiSelectTrigger, OptionList } from './multi-select-components'
+import { TriggerButton, DropdownContent } from './multi-select-helpers'
 
 export interface MultiSelectOption {
   value: string
@@ -44,7 +42,6 @@ export function MultiSelect({
       } else {
         onChange([...selected, value])
       }
-      // Fechar o dropdown após seleção
       setOpen(false)
     },
     [selected, onChange]
@@ -58,136 +55,17 @@ export function MultiSelect({
     return options.filter(option => selected.includes(option.value))
   }, [options, selected])
 
-  // Component ready for production
-
-  // TEMPORARY: Completely simplified version without Popover
   return (
     <div className="relative">
-      {/* Trigger Button */}
-      <button
-        type="button"
-        className="flex h-10 w-full items-center justify-between rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-        onClick={handleToggle}
-      >
-        <span className="flex-1 text-left">
-          {selectedOptions.length > 0 ? (
-            <div className="flex flex-wrap gap-1">
-              {selectedOptions.slice(0, maxCount).map(option => (
-                <span key={option.value} className="inline-flex items-center bg-secondary text-secondary-foreground px-2 py-0.5 rounded text-xs">
-                  {option.label}
-                  <button
-                    type="button"
-                    className="ml-1 hover:text-destructive"
-                    onClick={(e) => {
-                      e.stopPropagation()
-                      handleUnselect(option.value)
-                    }}
-                  >
-                    ×
-                  </button>
-                </span>
-              ))}
-              {selectedOptions.length > maxCount && (
-                <span className="text-xs text-muted-foreground">
-                  +{selectedOptions.length - maxCount} mais
-                </span>
-              )}
-            </div>
-          ) : (
-            <span className="text-muted-foreground">{placeholder}</span>
-          )}
-        </span>
-        <svg className="h-4 w-4 opacity-50" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-        </svg>
-      </button>
-
-      {/* Options List - Toggle visibility */}
-      {open && options.length > 0 && (
-        <div 
-          className="absolute top-full left-0 right-0 z-50 mt-1 max-h-[200px] overflow-y-auto rounded-md border border-border bg-popover shadow-lg"
-        >
-          {/* Search input */}
-          <div className="p-2 border-b border-border">
-            <input
-              type="text"
-              placeholder="Buscar..."
-              className="w-full px-2 py-1 text-sm border border-input rounded bg-background text-foreground placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring"
-              onChange={(e) => {
-                // TODO: Implement search functionality if needed
-                void e.target.value // Prevent unused parameter warning
-              }}
-            />
-          </div>
-        
-          {options.length === 0 ? (
-            <div className="p-2 text-sm text-muted-foreground">Nenhuma opção disponível</div>
-          ) : (
-            options.map(option => {
-              const isSelected = selected.includes(option.value)
-              return (
-                <div
-                  key={option.value}
-                  role="button"
-                  tabIndex={0}
-                  className={`flex items-center px-3 py-2 text-sm cursor-pointer hover:bg-accent hover:text-accent-foreground text-popover-foreground ${
-                    isSelected ? 'bg-accent/50' : ''
-                  }`}
-                  onClick={() => handleSelect(option.value)}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter' || e.key === ' ') {
-                      e.preventDefault()
-                      handleSelect(option.value)
-                    }
-                  }}
-                >
-                  <div className="mr-2 h-4 w-4 flex items-center justify-center">
-                    {isSelected ? (
-                      <svg className="h-4 w-4 text-primary" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
-                      </svg>
-                    ) : null}
-                  </div>
-                  {option.label}
-                </div>
-              )
-            })
-          )}
-        </div>
-      )}
+      <TriggerButton
+        open={open}
+        selectedOptions={selectedOptions}
+        maxCount={maxCount}
+        placeholder={placeholder}
+        onToggle={handleToggle}
+        onUnselect={handleUnselect}
+      />
+      <DropdownContent open={open} options={options} selected={selected} onSelect={handleSelect} />
     </div>
   )
-
-  // Original Popover implementation (commented for debugging)
-  /*
-  return (
-    <Popover open={open} onOpenChange={(newOpen) => {
-      console.log('🔍 Popover onOpenChange:', { from: open, to: newOpen })
-      setOpen(newOpen)
-    }}>
-      <PopoverTrigger asChild>
-        <MultiSelectTrigger
-          open={open}
-          selected={selected}
-          selectedOptions={selectedOptions}
-          maxCount={maxCount}
-          placeholder={placeholder}
-          className={className}
-          onToggle={handleToggle}
-          onUnselect={handleUnselect}
-        />
-      </PopoverTrigger>
-      <PopoverContent 
-        className="w-[200px] p-0 z-50" 
-        align="start"
-        style={{ backgroundColor: 'white', border: '1px solid #ccc', visibility: 'visible', display: 'block' }}
-      >
-        <div style={{ padding: '8px', backgroundColor: 'yellow', fontSize: '12px' }}>
-          🔍 DEBUG: PopoverContent renderizado!
-        </div>
-        <OptionList options={options} selected={selected} onSelect={handleSelect} />
-      </PopoverContent>
-    </Popover>
-  )
-  */
 }
